+++ Applications/Spire/Source/UI/MaxFloorSpinBox.cpp
@@ -83,6 +83,6 @@ void MaxFloorSpinBox::AdjustIncrement(KeyModifiers modifier) {
   auto quantityIncrement = m_userProfile->GetInteractionProperties().Get(
     *m_security).m_quantityIncrements[static_cast<int>(modifier)];
   if(quantityIncrement != singleStep()) {
-    setSingleStep(quantityIncrement);
+    setSingleStep(static_cast<int>(quantityIncrement));
   }
 }
